requirements.txt
python-telegram-bot==13.11

ephem~=4.1.3